fix dependencies / breakage due to restructuring
