+++ tools/storage-explorer/type.go
@@ -21,9 +21,9 @@ package main
 import (
 	"github.com/onflow/cadence"
 	jsoncdc "github.com/onflow/cadence/encoding/json"
-	"github.com/onflow/cadence/interpreter"
 	"github.com/onflow/cadence/runtime"
-	"github.com/onflow/cadence/sema"
+	"github.com/onflow/cadence/runtime/interpreter"
+	"github.com/onflow/cadence/runtime/sema"
 )
 
 func prepareType(value interpreter.Value, inter *interpreter.Interpreter) (result any, description string) {
